+++ azure-storage-blob/azure/storage/blob/_constants.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 
 __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
-__version__ = '2.0.1'
+__version__ = '2.1.0'
 
 # x-ms-version for storage service.
 X_MS_VERSION = '2019-02-02'