+++ azure-storage-blob/ChangeLog.md
@@ -3,9 +3,9 @@
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
 ## Version XX.XX.XX
-
 - Added Batch Delete Blob API
 - Added Batch Set Standard Blob Tier API(for BlockBlob)
+- Added Blob Tier support PutBlob/PutBlockList/CopyBlob APIs
 
 ## Version 2.0.1:
 - Updated dependency on azure-storage-common.